Add files via upload
As the name of my project (Payroll management system) .This is all about to generate salary slip of an employee. It is the the process by which employers pay wages to their employees. It's also how they demonstrate their commitment to their workers, fulfill their obligations to government agencies and keep financial records in order.
My project is based on certain conditions like employers pay wages to an employee on the basis of their salary ..HRA,TA,PF, etc are decided and gross salary will be generated.
It will generate the net  salary slip of an employee  on console and  also in a file named as record.txt  it will be save automatically.
So That is all about my java project.
Thankyou .